Summary:
copy_impl improvement:

Before: 1732 ns
After:.    1159 ns

remote_request_only

Before: Milliseconds per iter: 1.22971. Iters per second: 813.2
             0.169993 ms.    13.9123%. aten::to (155 nodes)

After:    Milliseconds per iter: 1.18376. Iters per second: 844.765
             0.119433 ms.    10.3746%. aten::to (155 nodes)

Test Plan: buck test caffe2:ATen-core-test

Differential Revision: D26967349